불필요 API 요청 최소화 & API 요청 시 요청간 휴식기 최소화

+++ src/main/java/my/trader/coin/util/TimeUtility.java
@@ -3,13 +3,17 @@ package my.trader.coin.util;
 import org.springframework.stereotype.Component;
 
 /**
- * 틱톡틱톡 시계는 와치 시계는 와치
+ * 틱톡틱톡 시계는 와치 시계는 와치.
  */
 @Component
 public class TimeUtility {
-  public static void sleep(long sec) {
+  /**
+   * 스레드는 자는중.
+   * @param sec 몇초동안 재울까요?
+   */
+  public static void sleep(double sec) {
     try {
-      Thread.sleep(sec * 1000);
+      Thread.sleep((long) (sec * 1000)); // 초를 밀리초로 변환
     } catch (InterruptedException e) {
       Thread.currentThread().interrupt();
     }
